+++ ball/st_edit.h
@@ -1,0 +1,8 @@
+#ifndef ST_EDIT_H
+#define ST_EDIT_H
+
+#include "state.h"
+
+extern struct state st_edit;
+
+#endif